Refactored to use component functions
