+++ src/App.tsx
@@ -4,8 +4,8 @@ import { createStore } from 'redux'
 import { Container } from 'semantic-ui-react'
 import 'semantic-ui-css/semantic.min.css';
 import './App.css';
-import ItemView from "./ItemView";
 import dbApp from "./State/Reducer";
+import MainView from './components/MainView';
 
 const store = createStore(dbApp,  (window as any).__REDUX_DEVTOOLS_EXTENSION__ && (window as any).__REDUX_DEVTOOLS_EXTENSION__());
 
@@ -16,7 +16,7 @@ class App extends Component {
         return (
             <Container>
                 <Provider store={store}>
-                    <ItemView/>
+                    <MainView/>
                 </Provider>
             </Container>
         );
